address some race conditions with Program shutdown and clearing thread-local data
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/qore/trunk@5761 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ lib/qore-main.cpp
@@ -66,7 +66,7 @@ static unsigned long q_openssl_id_function(void) {
 #if (defined _WIN32 || defined __WIN32__) && ! defined __CYGWIN__ 
    return GetCurrentThreadId();
 #else
-   return pthread_self();
+   return (unsigned long)pthread_self();
 #endif
 }
